+++ ext/JuMPModels/truck_trailer.jl
@@ -7,7 +7,7 @@ The Truck Trailer Problem:
 function OptimalControlProblems.truck_trailer(
     ::JuMPBackend;
     data::Array{Float64,2}=[0.4 0.1 0.2; 1.1 0.2 0.2; 0.8 0.1 0.2],
-    nh::Int64=200,
+    nh::Int=200,
 )
     # parameters
     if size(data) != (3, 3)
